[REFACTOR] Refactor package name to cdh47 with code formatting to checkstyle
