+++ cdh47/src/META-INF/services/org.pentaho.hadoop.shim.spi.HadoopShim
@@ -1,1 +1,1 @@
-org.pentaho.hadoop.shim.cdh42.HadoopShim
+org.pentaho.hadoop.shim.cdh47.HadoopShim
