+++ cdh47/src/META-INF/services/org.pentaho.hadoop.shim.spi.PigShim
@@ -1,1 +1,1 @@
-org.pentaho.hadoop.shim.cdh42.PigShim
+org.pentaho.hadoop.shim.cdh47.PigShim
