+++ cdh47/src/META-INF/services/org.pentaho.hadoop.shim.spi.SnappyShim
@@ -1,1 +1,1 @@
-org.pentaho.hadoop.shim.cdh42.SnappyShim
+org.pentaho.hadoop.shim.cdh47.SnappyShim
