+++ cdh47/src/META-INF/services/org.pentaho.hadoop.shim.spi.SqoopShim
@@ -1,1 +1,1 @@
-org.pentaho.hadoop.shim.cdh42.ClassPathModifyingSqoopShim
+org.pentaho.hadoop.shim.cdh47.ClassPathModifyingSqoopShim
