--- cdh47/src/org/pentaho/hadoop/shim/cdh42/HadoopShim.java
@@ -1,43 +1,0 @@
-/*******************************************************************************
-*
-* Pentaho Big Data
-*
-* Copyright (C) 2002-2013 by Pentaho : http://www.pentaho.com
-*
-*******************************************************************************
-*
-* Licensed under the Apache License, Version 2.0 (the "License");
-* you may not use this file except in compliance with
-* the License. You may obtain a copy of the License at
-*
-*    http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing, software
-* distributed under the License is distributed on an "AS IS" BASIS,
-* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-* See the License for the specific language governing permissions and
-* limitations under the License.
-*
-******************************************************************************/
-
-package org.pentaho.hadoop.shim.cdh42;
-
-
-import org.apache.hive.jdbc.HiveDriver;
-import org.pentaho.hadoop.shim.common.CommonHadoopShim;
-
-public class HadoopShim extends CommonHadoopShim {
-  
-  static {
-    JDBC_DRIVER_MAP.put("hive2",org.apache.hive.jdbc.HiveDriver.class); 
-  }
-  @Override
-  protected String getDefaultNamenodePort() {
-    return "8020";
-  }
-  
-  @Override
-  protected String getDefaultJobtrackerPort() {
-    return "8021";
-  }
-}
